hackportal/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.2
